Refactor to keep trainer.py clean
