+++ config/config_jepa.yml
@@ -66,6 +66,10 @@ forecast_att_dense_rate: 1.0
 with_step_conditioning: True # False
 
 healpix_level: 5
+# Use 2D RoPE instead of traditional global positional encoding
+# When True: uses 2D RoPE based on healpix cell coordinates (lat/lon)
+# When False: uses traditional pe_global positional encoding
+rope_2D: True
 
 with_mixed_precision: True
 with_flash_attention: True
